move integration tests to subfolder
