+++ tests/integration/searcher/suite/suite.go
